Add files via upload
The Google Snake Game WordPress Plugin is an engaging and interactive addition to any WordPress website, allowing users to enjoy the classic Snake game directly within posts or pages. This plugin is designed to offer a fun and nostalgic experience to website visitors, enhancing user engagement and providing a unique form of entertainment.